refactor: inline response cycle execution into composition node
Flatten abstraction by removing ResponseCycle.ts entirely:
- Inline flow creation/execution directly in ResponseCycleCompositionNode
- Update test to use createResponseCycleFlow() directly
- Delete src/agent/core/ResponseCycle.ts (no longer needed)
- Update @see references in ToolUseCycle.ts

Call path is now:
  ResponseCycleCompositionNode.exec()
    → createResponseCycleFlow()
    → flow.run()
