+++ src/agent/core/ToolUseCycle.ts
@@ -7,7 +7,7 @@
  * Supports reactive, session-based execution where the agent can call tools
  * and wait for user input.
  *
- * @see ResponseCycle for workflow-based cycle execution
+ * @see ResponseCycleFlow for workflow-based cycle execution
  */
 
 // Standard library imports
@@ -63,7 +63,7 @@ export interface ToolUseCycleResult {
  *
  * @param input - Cycle input with options, messages, and store
  * @returns Result with failedWithError and userCancelled flags
- * @see executeResponseCycleCore for workflow-based cycle execution
+ * @see ResponseCycleFlow for workflow-based cycle execution
  */
 export async function runToolUseCycle<C = unknown>(
   input: ToolUseCycleInput<C>,
