+++ src/test/agent/core/ResponseCycleBackground.test.ts
@@ -13,7 +13,12 @@ import {
   AgentType,
   AgentCategory,
 } from '@agent/core/AgentDataclass';
-import { executeResponseCycleCore } from '@agent/core/ResponseCycle';
+import {
+  createResponseCycleFlow,
+  type ResponseCycleShared,
+  type ResponseCycleState,
+} from '@agent/core/flows/ResponseCycleFlow';
+import { createRetryState } from '@agent/core/flows/RetryState';
 import { AgentWorkspaceState } from '@agent/core/AgentWorkspaceState';
 import { AgentExecutionContext } from '@agent/runtime/AgentExecutionContext';
 // Type imports
@@ -324,32 +329,50 @@ describe('ResponseCycle background reasoning logs', () => {
     const run = new AgentRunState();
     const workspace = AgentWorkspaceState.create();
     const userVars = {} as Record<string, any>;
+    const outputLocation = createWorkspaceLocation(
+      WorkspaceFS.fullPath('output.txt'),
+      'output.txt',
+    );
 
-    await executeResponseCycleCore({
-      options: {
-        modelHandler: handler,
-        agentSetting,
-        agentConfig,
-        agentPrompt,
-        userVars,
-        logger: loggerStub,
-        client: {} as OpenAI,
-        checkInterruption: () => false,
-        setAbortController: () => {},
-        context: new AgentExecutionContext({
-          streamId: 'test-stream' as StreamTabId,
-        }),
-        fileService: new TaskRunFileService(),
-      },
-      messages,
-      outputLocation: createWorkspaceLocation(
-        WorkspaceFS.fullPath('output.txt'),
-        'output.txt',
-      ),
+    // Create shared state for the cycle flow
+    const shared: ResponseCycleShared = {
+      state: {
+        messages,
+        outputLocation,
+        endTurn: false,
+        shouldStop: false,
+        outputExists: false,
+        systemPrompt: undefined,
+        debug: undefined,
+        responseObject: undefined,
+        responseTimeMs: undefined,
+        stopReason: undefined,
+        processedResponse: undefined,
+      } satisfies ResponseCycleState,
+      retryState: createRetryState(),
+    };
+
+    // Create and run the flow directly
+    const flow = createResponseCycleFlow();
+    flow.setServices({
+      modelHandler: handler,
+      agentSetting,
+      agentConfig,
+      agentPrompt,
+      userVars,
+      logger: loggerStub,
+      client: {} as OpenAI,
+      checkInterruption: () => false,
+      setAbortController: () => {},
+      context: new AgentExecutionContext({
+        streamId: 'test-stream' as StreamTabId,
+      }),
+      fileService: new TaskRunFileService(),
       round,
       run,
       workspace,
     });
+    await flow.run(shared);
 
     const thinkingLogs = loggedEvents.filter(
       (event) =>
